Resolved task5 from Matrix
+ isUnique func
+transposeIfMatrixHasNotEqualSumOfRows func
* getSum refactoring (int -> long long)

+++ libs/data_structures/matrix/matrix.c
@@ -268,8 +268,8 @@ matrix *createArrayOfMatrixFromArray(const int *values,
     return ms;
 }
 
-int getSum(int *a, int n) {
-    int sum = 0;
+long long getSum(int *a, int n) {
+    long long sum = 0;
     for (int i = 0; i < n; ++i)
         sum += a[i];
